Polish up the user interface.
This patch makes the smart refreshing of the table (such that rows stay open) into a reusable function which can be called for approval/removal/rescheduling events as well as browser auto refresh. It also removes highlighting when expanding a row and removed the ability to click highlight a subrow.

	modified:   productionsystem/webapp/templates/dashboard_template.html