Disable Client-Side sorting and don't limit number of displayed parametricjobs.
This is necessary since the client side will not naturally sort the "request#.parametricjob#" strings. Even if it could, they are already sorted server side. We also disable the limit on the number of displayed parametric jobs since we don't have the paging controls we need to see them all.

	modified:   productionsystem/webapp/templates/dashboard_template.html