Fixing Service timestamp update.
This required the manual updating of the timestamp as although onupdate hook works when the new value differs from the current DB value, it does effectively a no-op when they are the same.

This means that while can use the timestamp to give the length of time a service has been up, it was no good as a heartbeat as successive setting of status=UP will not touch the timestamp.

	modified:   productionsystem/monitoring/MonitoringDaemon.py
	modified:   productionsystem/sql/models/Services.py