Deletion done in monitoring loop.
This commit changes the behaviour such that the webapp no longer waits on the server doing the deletion. Now a request is simple marked as "Removing" and the monitoring loop does the deletion and DIRAC cleanup. This is done in chunks of 1000 Dirac jobs to avoid timeout issues and since spees isn't an issue for the monitoring loop.

This commit also makes requests marked as "Removed" disabled from user input (clicking with either left or right mouse buttons) as well as making them greyed out.

Finally this commit fixes a long-standing bug where subtable rows could be selected and right clicked on to get the context menu which doesn't make sence for them

This commit fixes https://github.com/alexanderrichards/ProductionSystem/issues/42 and fixes https://github.com/alexanderrichards/ProductionSystem/issues/4

	modified:   productionsystem/monitoring/MonitoringDaemon.py
	modified:   productionsystem/sql/enums.py
	modified:   productionsystem/sql/models/ParametricJobs.py
	modified:   productionsystem/sql/models/Requests.py
	modified:   productionsystem/webapp/services/RESTfulAPI.py
	modified:   productionsystem/webapp/templates/dashboard_template.html